Refactor project structure and update configurations: add .dockerignore, modify Dockerfile, adjust paths in server and client configurations, and enhance package management in lib.
