+++ .dockerignore
@@ -1,0 +1,7 @@
+node_modules
+npm-debug.log
+yarn-error.log
+.env
+.DS_Store
+.git
+.gitignore
