+++ client/webpack.config.ts
@@ -14,7 +14,7 @@ const config: WebpackConfiguration = {
 
     output: {
         filename: "bundle.js",
-        path: path.resolve(__dirname, "../public"),
+        path: path.resolve(__dirname, "../server/public"),
     },
 
     devServer: {
