+++ lib/tsconfig.json
@@ -5,7 +5,7 @@
         "declarationMap": true,
         "rootDir": ".",
         "composite": true,
-        "outDir": "dist",
+        "outDir": "dist"
     },
     "references": [] // * Any project that is referenced must itself have a `references` array (which may be empty).
 }
