--- public/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <title>Jackal</title>
-        <script defer src="/public/bundle.js"></script>
-        <link rel="icon" type="image/x-icon" href="/public/favicon.ico" />
-    </head>
-    <body>
-        <div id="root"></div>
-    </body>
-
-    <script>
-        if ("serviceWorker" in navigator) {
-            window.addEventListener("load", () => {
-                navigator.serviceWorker
-                    .register("service-worker.js")
-                    .then((registration) => {
-                        console.log("Service Worker registered: ", registration);
-                    })
-                    .catch((registrationError) => {
-                        console.error("Service Worker registration failed: ", registrationError);
-                    });
-            });
-        }
-    </script>
-</html>
