--- server/DockerFile
@@ -1,15 +1,0 @@
-FROM Node:20
-
-WORKDIR /usr/src/app
-
-COPY . .
-
-COPY ../public /usr/src/public
-COPY .env .env
-
-RUN npm install
-RUN npm run build
-
-EXPOSE 3000
-
-CMD ["npm", "start"]
