+++ server/package.json
@@ -31,6 +31,7 @@
     "license": "ISC",
     "dependencies": {
         "@types/color": "^4.2.0",
+        "axios": "^1.7.9",
         "colors": "^1.4.0",
         "dotenv": "^16.4.5",
         "express": "^4.21.1",
